Use layout file for saving and loading layout
Replace storing layout in a setting with a dedicated layout
file (layout.fyl). This allows the layout to be edited outside
of fooyin.

FyWidget subclasses should also override the layoutName
method to change the name written to file. An optional display name can
also be set when registering widgets in the widget factory for use in
menus.