Store Year/DATE frame as string
The TDRC frame stores data as a timestamp, which could be "YYYY-MM-dd"
or just "YYYY". Rather than converting to an integer (and only storing
the year), also store it as a string - date. This will be useful for the
tag editor later.

Also restructor the model classes so they are more readable.