Add queue to LibraryScanner
When scanning libraries, the event queue is processed to check for stop
signals. This can lead to a scanLibrary request whilst the thread is
still busy, which then returns. In order to resolve this, implement a
basic queue for scan requests, and process this once the thread is idle.

Also implement LibraryInfo as a struct, and add methods to rename an
existing library.