Store libraries seperately
Store each library as an individual instance. Create a library container so connected signals remain intact on library change. Also implement a unified library to provide an interface for accessing all libraries.

For the PlaylistWidget, store selected tracks in the widget itself, and
handle the library playlist interface.

On the settings side, store the SettingsDialogController in the
SettingsManager to simplify dependencies.

Additionally restructure Application, removing use of pimpl.