[gui] Implement custom comparator for QModelIndex
The default operator< implementation for QModelIndex works fine for lists and tables, but is not designed to handle trees with large amounts of nesting.

This implementation assumes a single column, and iterates up the tree hierarchy until a common ancestor is found (which may be root). The row number for each QModelIndex can then be correctly compared.