[gui] LibraryTree: Prevent duplicate nodes
Due to how the LibraryTree model is loaded in batches, nodes could be appended to their parent to be added by fetchMore more than once if there were tracks with the same parents across different batches, but those parents had yet to be added as they were not visible.
This is fixed by keeping track of each node until it has been properly added to the model.
Additionally, LibraryTreeItem::pending is now set to false by default so pending nodes no longer need to be set to true before being sent to the model in the populator.