[core,gui,utils] Rewrite readMetaData and cover management
Tagging: readMetaData now uses QMimeDatabase to determine file before instantiating a TagLib::File object of that type from a FileStream

Covers: CoverProvider has been moved to the gui library where it makes more sense. Cover thumbnails are no longer generated when reading metadata, but when a cover is first requested. This is handled asynchronously in a coroutine, with a signal being emitted when a track's cover is ready.