[gui] Rewrite playlist preset script parsing
Splits the track scripts into explicit left and right entries, and the subheaders into explicit levels of nesting. This makes it easier to parse the scripts, and allows subheaders to be used in function arguments.

Several functions in the playlist model have also been improved to account for these changes, and to fix incorrect behaviour when copying tracks across different playlists.