[gui,utils] Improve TrackSelection context handling
Copies the QActions from the track and playlist Utils::ActionContainers to preserve state changes. Utils::cloneMenu has been removed as it is no longer used and is incompatible with Utils::ActionManager.
Update action state of the above whenever the track selection changes. This will provide more reliable behaviour if shortcuts are used to execute actions.