Store track ids rather than tracks in mimedata
There was a not-uncommon issue with deserialising tracks during drag and drop operations. On top of this, tracks are implicitly shared and so serialising track lists is potentially a very expensive operation both int terms of performance and memory usage.

This is resolved by storing the track ids instead. The original TrackList can then be obtained through MusicLibrary::tracksForIds.