[core] Replace use of std::bit_cast
Audio data is stored in the AudioBuffer class as a vector of std::byte. In order to access the real underlying values, bit_cast was used to cast a pointer to the raw array of the vector.
This is only defined behaviour however when the type being cast to has the same size as the declared type (and ideally as a value). Otherwise, the cast violates strict aliasing and will potentially lead to alignment issues. The same issues would also occur when using reinterpret_cast.
To avoid the above issues, use std::memcpy to copy the required number of bytes into the correct type.