[gui] Playlist: Fix flickering vertical scrollbar when resizing
If stretch is enabled for the header view and artwork columns are present, resizing the playlist to hide the scrollbar will auto-resize the columns. This may cause the contents to exceed the viewport, making the scrollbar reappear and potentially creating an endless loop of showing and hiding the scrollbar.
Prevent this by ignoring the first resize column call after updating the scrollbars. There still exists some flickering when actively resizing, but only for a small range.

Fix #141